requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-environ
django-filter